data/crawlers/requirements.txt
beautifulsoup4==4.13.4
pymongo==4.12.1
python-dotenv==1.1.0
Requests==2.32.3
undetected_chromedriver==3.5.4
setuptools>=65.5.0 
